// src/api/job/message/contrat-type.message.ts
export enum JobCRUDMessage {
  // Créer
  CREATE_SUCCESS = 'Métier créé avec succès.',
  CREATE_ERROR = 'Erreur lors de la création du métier.',

  // Lire
  READ_SUCCESS = 'Métier récupéré avec succès.',
  READ_ERROR = 'Erreur lors de la récupération du métier.',

  // Mettre à jour
  UPDATE_SUCCESS = 'Métier mis à jour avec succès.',
  UPDATE_ERROR = 'Erreur lors de la mise à jour du métier.',

  // Supprimer
  DELETE_SUCCESS = 'Métier supprimé avec succès.',
  DELETE_ERROR = 'Erreur lors de la suppression du métier.',
}